+++ packages/integration-sdk-core/src/types/jobState.ts
@@ -4,6 +4,11 @@ export interface GraphObjectFilter {
   _type: string;
 }
 
+export interface GraphObjectLookupKey {
+  _type: string;
+  _key: string;
+}
+
 export type GraphObjectIteratee<T> = (obj: T) => void | Promise<void>;
 
 /**
@@ -63,6 +68,12 @@ export interface JobState {
    * to add a single relationship to the collection.
    */
   addRelationships: (relationships: Relationship[]) => Promise<void>;
+
+  /**
+   * Gets an entity by _key and _type
+   * Throws when !== 1 entity
+   */
+  getEntity: (lookupKey: GraphObjectLookupKey) => Promise<Entity>
 
   /**
    * Allows a step to iterate all entities collected into the job state, limited