+++ packages/integration-sdk/CHANGELOG.md
@@ -9,10 +9,19 @@ and this project adheres to
 
 ## Unreleased
 
+### Added
+
+- Added `getEntity({ _type, _key })` function
+
 ## 2.5.0 - 2020-07-28
+
+### Added
+
+- `j1-integration visualize` added mapped relationships as dashed lines. 
 
 ### Changed
 
+- `j1-integration visualize` entities are colored by `_type`.
 - Updated `IntegrationConfig` to support asynchronous `getStepStartStates`. See
   [#254](https://github.com/JupiterOne/sdk/issues/254) for more information.